+++ packages/datasource-replica/src/sequelize.ts
@@ -92,6 +92,7 @@ export async function createSequelize(logger: Logger, options: ReplicaDataSource
   const sequelize = await buildSequelizeInstance(options.cacheInto, logger, {
     tables: [],
     version: 1,
+    source: '@forestadmin/datasource-sql',
   });
 
   // This table should never need to change => use normal sync