+++ packages/forest-cloud/test/commands/publish.test.ts
@@ -30,7 +30,7 @@ describe('publish command', () => {
 
     expect(cmd.outputs).toEqual([
       cmd.spinner.start('Publishing code customizations'),
-      cmd.spinner.start('Publishing code customizations (operation cannot be cancelled)'),
+      cmd.spinner.start('Publishing code customizations'),
       cmd.spinner.succeed('Code customizations published'),
       cmd.spinner.stop(),
     ]);
@@ -53,6 +53,10 @@ describe('publish command', () => {
         cmd.spinner.start('Publishing code customizations'),
         cmd.spinner.warn('There is already deployed customization code on your project'),
         cmd.spinner.info('Last code pushed yesterday, by John Doe (johndoad@forestadmin.com)'),
+        cmd.spinner.info(
+          // eslint-disable-next-line max-len
+          'For the next time, you can publish your customizations with the --force option to skip this step',
+        ),
         cmd.spinner.stop(),
         cmd.question('Do you really want to overwrite these customizations? (yes/no) '),
         cmd.spinner.fail('Operation aborted'),
@@ -72,7 +76,7 @@ describe('publish command', () => {
 
         expect(cmd.outputs).toEqual([
           cmd.spinner.start('Publishing code customizations'),
-          cmd.spinner.start('Publishing code customizations (operation cannot be cancelled)'),
+          cmd.spinner.start('Publishing code customizations'),
           cmd.spinner.succeed('Code customizations published'),
           cmd.spinner.stop(),
         ]);
@@ -91,7 +95,7 @@ describe('publish command', () => {
 
       expect(cmd.outputs).toEqual([
         cmd.spinner.start('Publishing code customizations'),
-        cmd.spinner.start('Publishing code customizations (operation cannot be cancelled)'),
+        cmd.spinner.start('Publishing code customizations'),
         cmd.spinner.fail('Something went wrong: An error occurred'),
         cmd.spinner.stop(),
       ]);
@@ -109,7 +113,7 @@ describe('publish command', () => {
 
       expect(cmd.outputs).toEqual([
         cmd.spinner.start('Publishing code customizations'),
-        cmd.spinner.start('Publishing code customizations (operation cannot be cancelled)'),
+        cmd.spinner.start('Publishing code customizations'),
         cmd.spinner.fail('An error occurred'),
         cmd.spinner.stop(),
       ]);