+++ packages/plugin-flattener/test/index.test.ts
@@ -1,4 +1,4 @@
-import { flattenColumn, flattenRelation } from '../src/index';
+import { flattenColumn, flattenJsonColumn, flattenRelation } from '../src/index';
 
 describe('index', () => {
   test('should export flattenColumn', () => {
@@ -7,5 +7,9 @@ describe('index', () => {
 
   test('should export flattenRelation', () => {
     expect(flattenRelation).toBeInstanceOf(Function);
+  });
+
+  test('should export flattenJsonColumn', () => {
+    expect(flattenJsonColumn).toBeInstanceOf(Function);
   });
 });