+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,45 @@
+## @forestadmin/agent [1.38.7](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.38.6...@forestadmin/agent@1.38.7) (2024-03-14)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.43.6
+* **@forestadmin/datasource-toolkit:** upgraded to 1.32.2
+* **@forestadmin/forestadmin-client:** upgraded to 1.25.12
+
+## @forestadmin/agent [1.38.6](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.38.5...@forestadmin/agent@1.38.6) (2024-03-12)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.43.5
+
+## @forestadmin/agent [1.38.5](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.38.4...@forestadmin/agent@1.38.5) (2024-03-11)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.43.4
+
+## @forestadmin/agent [1.38.4](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.38.3...@forestadmin/agent@1.38.4) (2024-03-08)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/forestadmin-client:** upgraded to 1.25.11
+
 ## @forestadmin/agent [1.38.3](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.38.2...@forestadmin/agent@1.38.3) (2024-02-28)